LayoutTests/imported/w3c/web-platform-tests/workers/WorkerGlobalScope_ErrorEvent_filename-expected.txt

CONSOLE MESSAGE: line 3: Error: Error Message

PASS  WorkerGlobalScope onerror event handler argument: location  
